Remove ExpressionTyped
